Oh jesus fucking christ.
Fucking git and its shitty feedback caused me to miss the extra files the first time.
My anger at that caused me to copy a file into the wrong place. This.
Which also didn't undo the thing I meant to undo... so that got checked straight in.

So: nevermind any improvements I made. I'm over here fighting with the
version control software.

Dear software developers: any sufficiently confusing feature is indistiguishable
from a bug.

Git is a step forward in so many ways, but usability and sensible defaults are not among them.