Added a hack to improve the behavior of chording keys with META on Mac.
It's a shit situation. This would appear to be an Apple bug that prevents
dispatching a keyUp for a regular keyboard key when one of the META keys
is held down. It can be fixed by overriding sendEvent() in NSApplication,
but that's not something we can do, AFAIK, because Java is the application.

This hack improves the behavior in practice, but does not fix it.